requirements.txt
streamlit
mysql-connector-python
pandas
numpy

pip install -r requirements.txt